fix: float-parse BOX_HTTP_TIMEOUT, harden deadline guards + tests (review)
From the xhigh review of this PR:

- _http_timeout parsed with int(), so a fractional BOX_HTTP_TIMEOUT=1.5 hit
  ValueError and silently fell back to the LARGER default 30 — the opposite of
  the knob's purpose. Parse as float (mirroring BOX_SCAN_DEADLINE), reject
  non-finite/non-positive. Widen the client timeout hints to float.
- _scan_deadline's override path lacked the math.isfinite guard the env path
  got, so a nan override silently disabled the deadline. Unify both paths
  through the guard.
- test_scan_deadline_cuts_scan_short used a clock base of 0.0, so elapsed ==
  monotonic() and it couldn't catch a missing `- start`. Use a non-zero base.
- Add an end-to-end test that BOX_HTTP_TIMEOUT reaches the client's httpx
  timeout, and float/nan/inf parse cases.
- conftest: clear ambient BOX_SCAN_DEADLINE / BOX_HTTP_TIMEOUT so a shell value
  can't silently cut scan tests short.
- Trim the most redundant of the repeated deadline-rationale comments.

+++ boxadm_mcp/client.py
@@ -329,7 +329,7 @@ class BoxClient(_FolderReadMixin):
         enterprise_id: str,
         *,
         api_base: str = DEFAULT_API_BASE,
-        timeout: int = DEFAULT_TIMEOUT,
+        timeout: float = DEFAULT_TIMEOUT,
     ):
         self._client_id = client_id
         self._client_secret = client_secret
@@ -422,7 +422,7 @@ class BoxOAuthClient(_FolderReadMixin):
         *,
         token_cache: str | None = None,
         api_base: str = DEFAULT_API_BASE,
-        timeout: int = DEFAULT_TIMEOUT,
+        timeout: float = DEFAULT_TIMEOUT,
     ):
         self._client_id = client_id
         self._client_secret = client_secret
